# openapi.yaml
openapi: 3.1.0
info:
  title: ClassnoteX API
  description: Backend API for ClassnoteX - AI-powered recording and transcription
    app.
  version: 0.4.0
paths:
...
  /sessions:
    post:
      tags:
      - Sessions
      summary: Create Session
      operationId: create_session_sessions_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Sessions
      summary: List Sessions
      operationId: list_sessions_sessions_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: From Date
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: To Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
                title: Response List Sessions Sessions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /search/sessions:
    get:
      tags:
      - Search
      - Search
      summary: Search Sessions
      description: 'Search sessions with filters.

        Note: Firestore requires composite indexes for complex filtering.'
      operationId: search_sessions_search_sessions_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Query text (title match or simple search)
          title: Q
        description: Query text (title match or simple search)
      - name: mode
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mode
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: From Date
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: To Date
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
                title: Response Search Sessions Search Sessions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /search/tasks:
    get:
      tags:
      - Search
      - Search
      summary: Search Tasks
      description: 'Search tasks assigned to or owned by user.

        Assumes `tasks` root collection exists with `userId` or `assignee` fields.'
      operationId: search_tasks_search_tasks_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
                title: Response Search Tasks Search Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /search/decisions:
    get:
      tags:
      - Search
      - Search
      summary: Search Decisions
      description: "Search decisions extracted from sessions.\nAssumes `decisions`\
        \ root collection? Or `summaries` collection query?\nUser spec: \"3.4 summaries\
        \ / tasks / decisions\". \nUsually decisions are inside `summary` object or\
        \ separate.\nIf separate `decisions` not strictly defined as root collection,\
        \ we might check `summaries`.\nUser Spec 4.6 says: GET /search/decisions.\n\
        Let's assume we maintain a `decisions` collection or query Summaries.\nFor\
        \ MVP, let's query `sessions` and extract decisions? No, that's heavy.\nLet's\
        \ assume we query 'summaries' collection if it exists?\nSchema 3.4 says `summaries/{sessionId}`.\
        \ It's 1-to-1 with session? Or root collection `summaries`?\nIf root `summaries`\
        \ collection exists, we can query it."
      operationId: search_decisions_search_decisions_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DecisionResponse'
                title: Response Search Decisions Search Decisions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions/{session_id}/reaction:
    get:
      tags:
      - Reactions
      summary: Get Reaction State
      operationId: get_reaction_state_sessions__session_id__reaction_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionStateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Reactions
      summary: Set Reaction
      operationId: set_reaction_sessions__session_id__reaction_put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetReactionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionStateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateSessionRequest:
      properties:
        title:
          type: string
          title: Title
        mode:
          type: string
          title: Mode
          default: lecture
        userId:
          anyOf:
          - type: string
          - type: 'null'
          title: Userid
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        startAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Startat
        endAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Endat
        syncToGoogleCalendar:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Synctogooglecalendar
          default: false
        visibility:
          type: string
          title: Visibility
          default: private
        transcriptionMode:
          anyOf:
          - $ref: '#/components/schemas/TranscriptionMode'
          - type: 'null'
          default: cloud_google
      type: object
      required:
      - title
      title: CreateSessionRequest
    DecisionResponse:
      properties:
        id:
          type: string
          title: Id
        sessionId:
          type: string
          title: Sessionid
        content:
          type: string
          title: Content
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
      type: object
      required:
      - id
      - sessionId
      - content
      title: DecisionResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReactionStateResponse:
      properties:
        myEmoji:
          anyOf:
          - type: string
          - type: 'null'
          title: Myemoji
        counts:
          additionalProperties:
            type: integer
          type: object
          title: Counts
          default: {}
      type: object
      title: ReactionStateResponse
...
    SessionResponse:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        mode:
          type: string
          title: Mode
        userId:
          type: string
          title: Userid
        status:
          type: string
          title: Status
        createdAt:
          type: string
          format: date-time
          title: Createdat
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        isOwner:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Isowner
        sharedWithCount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sharedwithcount
        sharedUserIds:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Shareduserids
          default: []
        reactionCounts:
          anyOf:
          - additionalProperties:
              type: integer
            type: object
          - type: 'null'
          title: Reactioncounts
          default: {}
        ownerUserId:
          anyOf:
          - type: string
          - type: 'null'
          title: Owneruserid
        participantUserIds:
          items:
            type: string
          type: array
          title: Participantuserids
          default: []
        participants:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Participants
        visibility:
          type: string
          title: Visibility
          default: private
        autoTags:
          items:
            type: string
          type: array
          title: Autotags
          default: []
        topicSummary:
          anyOf:
          - type: string
          - type: 'null'
          title: Topicsummary
        summaryStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Summarystatus
          default: not_started
        quizStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Quizstatus
          default: not_started
        explainStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Explainstatus
          default: not_started
        diarizationStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Diarizationstatus
          default: not_started
        highlightsStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Highlightsstatus
          default: not_started
        isArchived:
          type: boolean
          title: Isarchived
          default: false
        lastOpenedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Lastopenedat
        startedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Startedat
        endedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Endedat
        durationSec:
          anyOf:
          - type: number
          - type: 'null'
          title: Durationsec
        hasTranscript:
          type: boolean
          title: Hastranscript
          default: false
        cloudTicket:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Cloudticket
        cloudAllowedUntil:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cloudalloweduntil
        cloudStatus:
          anyOf:
          - type: string
            enum: ["none", "allowed", "limited", "blocked"]
          - type: 'null'
          title: Cloudstatus
      type: object
      required:
      - id
      - title
      - mode
      - userId
      - status
      - createdAt
      title: SessionResponse
    SetReactionRequest:
      properties:
        emoji:
          anyOf:
          - type: string
            enum:
            - 🔥
            - 👏
            - 😇
            - 🤯
            - 🫶
          - type: 'null'
          title: Emoji
      type: object
      title: SetReactionRequest
...
    TaskResponse:
      properties:
        id:
          type: string
          title: Id
        sessionId:
          type: string
          title: Sessionid
        userId:
          type: string
          title: Userid
        title:
          type: string
          title: Title
        assignee:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignee
        dueDate:
          anyOf:
          - type: string
          - type: 'null'
          title: Duedate
        status:
          type: string
          title: Status
          default: open
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
          default: ai
      type: object
      required:
      - id
      - sessionId
      - userId
      - title
      title: TaskResponse
...
    TranscriptionMode:
      type: string
      enum:
      - cloud_google
      - device_sherpa
      - device_apple
      - dual_cloud_and_device
      title: TranscriptionMode
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token